split logging tests out
